fonts: Add fallback variant of Roboto
Harmony is missing some glyphs that users have come to expect on Android,
so use Roboto as a fallback to fix excessive tofu in some text.

Because Harmony's name has been replaced with "Roboto" for compatibility,
the fallback variants are named "Roboto Fallback".

Example of text with missing glyphs in Harmony: ⅙⅐⅛⅑⅒¹½⅓¼⅕

Change-Id: I11dd02966b34abf145dae3cc97913f79c9850744
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>